+++ Package@swift-5.3.swift
@@ -35,11 +35,14 @@ let package = Package(
             name: "AppCenter",
             path: "AppCenter/AppCenter",
             exclude: ["Support"],
-            cSettings: [
-                .define("APP_CENTER_C_VERSION", to:"\"4.4.2\""),
-                .define("APP_CENTER_C_BUILD", to:"\"1\""),
-                .headerSearchPath("**"),
-            ],
+            cSettings: {
+                var settings: [CSetting] = [
+                    .define("APP_CENTER_C_VERSION", to: "\"4.4.2\""),
+                    .define("APP_CENTER_C_BUILD", to: "\"1\"")
+                ]
+                settings.append(contentsOf: cHeaderSearchPaths)
+                return settings
+            }(),
             linkerSettings: [
                 .linkedLibrary("z"),
                 .linkedLibrary("sqlite3"),
@@ -55,10 +58,7 @@ let package = Package(
             dependencies: ["AppCenter"],
             path: "AppCenterAnalytics/AppCenterAnalytics",
             exclude: ["Support"],
-            cSettings: [
-                .headerSearchPath("**"),
-                .headerSearchPath("../../AppCenter/AppCenter/**"),
-            ],
+            cSettings: cHeaderSearchPaths,
             linkerSettings: [
                 .linkedFramework("Foundation"),
                 .linkedFramework("UIKit", .when(platforms: [.iOS, .tvOS])),
@@ -73,10 +73,7 @@ let package = Package(
             ],
             path: "AppCenterCrashes/AppCenterCrashes",
             exclude: ["Support", "Internals/MSACCrashesBufferedLog.hpp"],
-            cSettings: [
-                .headerSearchPath("**"),
-                .headerSearchPath("../../AppCenter/AppCenter/**"),
-            ],
+            cSettings: cHeaderSearchPaths,
             linkerSettings: [
                 .linkedFramework("Foundation"),
                 .linkedFramework("UIKit", .when(platforms: [.iOS, .tvOS])),
@@ -88,10 +85,7 @@ let package = Package(
             dependencies: ["AppCenter"],
             path: "AppCenterDistribute/AppCenterDistribute",
             exclude: ["Support"],
-            cSettings: [
-                .headerSearchPath("**"),
-                .headerSearchPath("../../AppCenter/AppCenter/**"),
-            ],
+            cSettings: cHeaderSearchPaths,
             linkerSettings: [
                 .linkedFramework("Foundation"),
                 .linkedFramework("SafariServices", .when(platforms: [.iOS])),
@@ -101,3 +95,47 @@ let package = Package(
         )
     ]
 )
+
+let projectHeaderSearchPaths = [
+    "../../AppCenter/AppCenter/Internals",
+    "../../AppCenter/AppCenter/Internals/Channel",
+    "../../AppCenter/AppCenter/Internals/Context/Device",
+    "../../AppCenter/AppCenter/Internals/Context/Session",
+    "../../AppCenter/AppCenter/Internals/Context/UserId",
+    "../../AppCenter/AppCenter/Internals/DelegateForwarder",
+    "../../AppCenter/AppCenter/Internals/HttpClient",
+    "../../AppCenter/AppCenter/Internals/HttpClient/Util",
+    "../../AppCenter/AppCenter/Internals/Ingestion",
+    "../../AppCenter/AppCenter/Internals/Ingestion/Util",
+    "../../AppCenter/AppCenter/Internals/Model",
+    "../../AppCenter/AppCenter/Internals/Model/CommonSchema",
+    "../../AppCenter/AppCenter/Internals/Model/Properties",
+    "../../AppCenter/AppCenter/Internals/Model/Util",
+    "../../AppCenter/AppCenter/Internals/Storage",
+    "../../AppCenter/AppCenter/Internals/Util",
+    "../../AppCenter/AppCenter/Internals/Vendor/Reachability",
+    "../../AppCenter/AppCenter/include",
+    "../../AppCenter/AppCenter/Model",
+    "../../AppCenterAnalytics/AppCenterAnalytics/include",
+    "../../AppCenterAnalytics/AppCenterAnalytics/Internals",
+    "../../AppCenterAnalytics/AppCenterAnalytics/Internals/Model",
+    "../../AppCenterAnalytics/AppCenterAnalytics/Internals/Session",
+    "../../AppCenterAnalytics/AppCenterAnalytics/Internals/Util",
+    "../../AppCenterAnalytics/AppCenterAnalytics/Model",
+    "../../AppCenterAnalytics/AppCenterAnalytics/TransmissionTarget",
+    "../../AppCenterCrashes/AppCenterCrashes/Internals",
+    "../../AppCenterCrashes/AppCenterCrashes/Internals/Model",
+    "../../AppCenterCrashes/AppCenterCrashes/Internals/Util",
+    "../../AppCenterCrashes/AppCenterCrashes/include",
+    "../../AppCenterCrashes/AppCenterCrashes/Model",
+    "../../AppCenterCrashes/AppCenterCrashes/WrapperSDKUtilities",
+    "../../AppCenterDistribute/AppCenterDistribute/Internals",
+    "../../AppCenterDistribute/AppCenterDistribute/Internals/Channel",
+    "../../AppCenterDistribute/AppCenterDistribute/Internals/Model",
+    "../../AppCenterDistribute/AppCenterDistribute/Internals/Version",
+    "../../AppCenterDistribute/AppCenterDistribute/Internals/Util",
+    "../../AppCenterDistribute/AppCenterDistribute/include",
+    "../../AppCenterDistribute/AppCenterDistribute/Model"
+]
+
+let cHeaderSearchPaths: [CSetting] = projectHeaderSearchPaths.map { .headerSearchPath($0) }